requirements.txt
python-telegram-bot==13.7
Telethon==1.21.1
asyncio==3.4.3
pytz==2021.1